api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
httptools==0.7.1
hydra-core==1.3.2
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
omegaconf==2.3.0
packaging==26.0
pillow==12.1.1
portalocker==3.2.0
pycocotools==2.0.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
setuptools==82.0.1
starlette==0.52.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
eva-decord==0.6.1
opencv-python==4.13.0.92
